feat(cascade): S3+S4 — unify on Matter.js, delete Rapier, normalized unit system (#1608 #1609)
S3: Delete Rapier WASM path entirely and promote the Matter.js engine to the
single unified engine used on all platforms (web + native).

- engine.ts: rewritten as the unified Matter.js engine (was Rapier/WASM)
- engine.native.ts: deleted — native now falls back to engine.ts
- __mocks__/@dimforge/rapier2d-compat.ts: deleted
- engine.test.ts, engineSimulation.test.ts, physics-parity.test.ts: deleted
- engine.native.test.ts → engine.unified.test.ts (import updated to ../engine)
- dropPhysics.test.ts: import updated from engine.native → engine
- package.json: remove @dimforge/rapier2d-compat dep + @dimforge transform pattern

S4: Normalized unit system — remove Rapier coordinate-space constants.

- engine.shared.ts: remove SCALE, GRAVITY_Y, RAPIER_SOLVER_ITERATIONS
- engine.shared.ts: add MATTER_GRAVITY_Y = 1.8 as a shared constant
  (equivalent to Rapier GRAVITY_Y=18 × SCALE=0.01 × 1/SCALE at 60 Hz)
- engine.ts: use MATTER_GRAVITY_Y from shared instead of local const

All platforms now run a single pixel-space physics world with no WASM
dependency, no coordinate-scale indirection, and no dual-engine maintenance.
